lab_guiado/Docker_K8s/aula8/03-lab_api_redis_worker/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
redis==5.0.8